Updated recipes to new format
> updated recipes to new format
> updated JEI to new format
